\name{A.data}
\docType{data}
\alias{A.data}
\title{Dose response data for rituximab and doxurubicin.}
\description{
  This data set includes absorbance measurements from two dose response experiments on rituximab and doxurubicin including five time points.
}
\usage{A.data}
\format{A list containg all information regarding the the setup of the experiments and the results.}
\references{
 Steffen Falgreen 2013
}
\keyword{datasets}