\name{bootstrap}
\alias{bootstrap}
\alias{bootstrapShiny}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create bootstrapped datasets based on the dose response experiments.
}
\description{
In order to achieve confidence intervals for summary statistics it is necessary to bootstrap the data. The bootstrapped datasets are normalised using the function \code{bgModel}. Since this will take some time we recommend only using 50 bootstrap samples.    
}
\usage{
bootstrap(A.data, update = TRUE, n.samples = 50, max.iter = 100, 
          type = c("parametric", "residual", "nonparametric"),  
          progressbar = "text", verbose = FALSE, save = TRUE, 
          shiny.input = NULL, session = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{A.data}{
An A.data object created by the function \code{bgModel}.
}
  \item{update}{
Should the analysis be updated or run from scratch. When set to \code{TRUE} only new experiments will be normalised. When \code{FALSE} The normalisation is conducted for all experiments.
}
  \item{n.samples}{
Numeric vslue indicating the number of bootstrap samples to create. Defeaults to 50.
}
  \item{max.iter}{
Numeric value indicating the maximum number of iterations allowed in order to obtain \code{n.samples} If for some reason the bootstrap procedure fails to create a data set, this value indicates the maximum number of iterations the precedure can be run. Should be at larger than or equal to \code{n.samples}. Defeaults to 100.
} 
  \item{type}{
Character indicating the bootstrap procedure to use. Current implimentation includes  \code{parametric} and \code{residual}.
}
  \item{progressbar}{
The type of progress bar used to show how far along the function is. Can be either \code{"window"}, \code{text} or \code{none}.
}

\item{verbose}{
Should the procedure indicate what bootstrap iteration it is currently working on.
}
\item{save}{
Should the data be saved.
}
  \item{shiny.input}{
Used for the shiny server.
}
  \item{session}{
Used for the shiny server.
}
}
\details{
It is possible to choose between parametric bootstrap and residual bootstrap. In parametric bootstrap the model fitted to the data is used generate new absorbance values. In residual bootstrap the absorbance values are generated by sampling the residuals of the fitted model.
}
\value{
The ouput of the function is an A.data object of class \code{bgModel}. This is a list with the following components:
 \item{meta.list}{This is a list of meta data objects.}
 \item{call}{A list containing information regarding the call to the function.}
 \item{auxiliary}{List of auxiliary data used by other functions.}
 \item{data}{List of data frames. The pre-processed bootstrapped data is stored in the element \code{bs.mean} columns BS:1,..., BS:\code{n.samples}.}
 \item{drug.color.correct}{Contains the results of the fitted dose reponse experiments for colour correction.}
 \item{fits}{List of the fitted objects. The fitted objects for model corrections is stored in element \code{bgModel}}
}
\references{
Steffen Falgreen et al. Exposure time independent summary statistics for assessment of drug dependent cell line growth inhibition (2013)
}
\author{
The function was written at department of haematology, Aalborg University Hospital and maintained by Steffen Falgreen.
}
\note{
After the data have been read into R the need to be preprocessed. If any drug colour correction plates have been established continue to the funktion \code{\link{drugColorCorrection}} otherwise continue to the function \code{\link{bgModel}}.
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{createMetaData}},\code{\link{readDBFData}},\code{\link{bgModel}},\code{\link{plotdrugColorCorrection}}
}
\examples{
require(DoseR)
## load Dose Response data
data(A.data)

A.data <- bootstrap(A.data    = A.data, 
                    n.samples = 50, 
                    max.iter  = 60,
                    type      = "parametric")

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{bootstrap}
