% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{cD.DM}
\alias{cD.DM}
\title{cD.DM Function to convert count dat object to LME4 input.}
\usage{
cD.DM(cD, weightByLL = TRUE, includeGroupStructure = FALSE)
}
\arguments{
\item{cD}{A countDat object containing FPKM values and at least one annotation column.}

\item{weightByLL}{Logical, weigh by log likelihood score.}

\item{includeGroupStructure}{Logical, include group structure.}
}
\value{
Returns LME4 input.
}
\description{
This function generates LME4 input.
}
\details{
This function converts count dat object.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
