% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChimericScaffold.R
\name{countCpt}
\alias{countCpt}
\title{countCpt Function to calculate countCpt.}
\usage{
countCpt(temp.cpt)
}
\arguments{
\item{temp.cpt}{temp.cpt file.}
}
\value{
Returns countCpt value for samples.
}
\description{
This function calculates countCpt.
}
\details{
This function calculates countCpt.
}
\examples{
data(G.RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
