% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{class}
\name{countDat-class}
\alias{countDat-class}
\title{An S4 class that stores count data.}
\description{
An S4 class that stores count data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Contains counts data.}

\item{\code{RPKM}}{Contains RPKM data.}

\item{\code{annotation}}{Contains annotation column data.}

\item{\code{replicates}}{Contains replicates data.}

\item{\code{rowObservables}}{Contains rowObservables data, (including seglens).}

\item{\code{sampleObservables}}{Contains sampleObservables data.}

\item{\code{orderings}}{Contains orderings data.}

\item{\code{nullPosts}}{Contains nullPosts data.}

\item{\code{cellObservables}}{Contains cellObservables data.}
}}

