% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMedianReadCounts.R
\name{countMedian}
\alias{countMedian}
\title{countMedian Function to calculate median.}
\usage{
countMedian(scaffolds, Sample, G_RESULT2)
}
\arguments{
\item{scaffolds}{Choice of scaffold.}

\item{Sample}{Choice of sample.}

\item{G_RESULT2}{G_RESULT file.}
}
\value{
Returns median value.
}
\description{
This function generates "median".
}
\details{
This function returns median expression for selected scaffold and sample.
}
\examples{
data(G.RESULT)
countMedian(scaffolds='DPSCF300112', Sample=1, G_RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
