% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChimericScaffold.R
\name{countReadsMara}
\alias{countReadsMara}
\title{countReadsMara Internal function to find chimeric scaffold.}
\usage{
countReadsMara(scaffolds, G_RESULT2)
}
\arguments{
\item{scaffolds}{Choice of scaffold.}

\item{G_RESULT2}{G_RESULT file.}
}
\value{
Returns countReadsMara data.
}
\description{
This function finds chimeric scaffold from underlying data.
}
\details{
This function returns data from internal function to find chimeric scaffold.
}
\examples{
data(G.RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
