% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldMeanRatios.R
\name{countReadsMara2}
\alias{countReadsMara2}
\title{countReadsMara2 Internal function to calculate scaffoldMeanRatios.}
\usage{
countReadsMara2(scaffolds, G_RESULT2)
}
\arguments{
\item{scaffolds}{Choice of scaffold.}

\item{G_RESULT2}{G_RESULT file.}
}
\value{
Returns scaffoldMeanRatios data.
}
\description{
This function generates scaffoldMeanRatios underlying data.
}
\details{
This function returns data from internal function to calculate scaffoldMeanRatios.
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
