% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafsFilter.R
\name{dafsFilter}
\alias{dafsFilter}
\title{dafsFilter	Function to filter expression data within a countDat object.}
\usage{
dafsFilter(cD, PLOT=TRUE)
}
\arguments{
\item{cD}{A countDat object.}

\item{PLOT}{Boolean, toggles plotting.}
}
\value{
Returns an invisible, filtered countDat object.
}
\description{
This function filters the expression of the supplied countDat object, by invoking the dafsFilter (dafs) function. dafsFilter is a filtering function used to remove rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countDat object using a Data Adaptive Flag filter. The internal function (dafs) uses a vector to store Kolmogorov Smirnov distance statistics, loops through cuts of the data to determine targeted K-S statistic, selects data greater than a quantile and runs Mclust on that data to determine theoretical distribution. The wrapper uses simpleFilter to determine first left-most local minima (using the Earth library).
}
\references{
BMC Bioinformatics, 2014, 15:92
}
\author{
AJ Vaestermark, JR Walters.
}
