% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{[,countDat,ANY,ANY,ANY-method}
\alias{[,countDat,ANY,ANY,ANY-method}
\title{subsetting method for countDat class}
\usage{
\S4method{[}{countDat,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{countDat object Value}

\item{i}{first dimension, subsetting parameter}

\item{j}{second dimension, subsetting parameter}

\item{...}{Passthrough arguments.}

\item{drop}{Value, Logical.}
}
\value{
subsetted object
}
\description{
subsetting method for countDat class
}
