% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChimericScaffold.R
\name{findChimericScaffold}
\alias{findChimericScaffold}
\title{findChimericScaffold Wrapper function to find chimeric scaffold.}
\usage{
findChimericScaffold(G__RESULT)
}
\arguments{
\item{G__RESULT}{G_RESULT file.}
}
\value{
Returns findChimericScaffold value for samples.
}
\description{
This function finds chimeric scaffold.
}
\details{
This function finds chimeric scaffold.
}
\examples{
data(G.RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
