% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.RESULT-data.R
\docType{data}
\name{G_RESULT}
\alias{G_RESULT}
\title{G_RESULT data set}
\format{An object of class \code{CompressedGRangesList} of length 5397.}
\usage{
data("G.RESULT")
}
\description{
G_RESULT data set
}
\references{
The "Zwyx" package, 2018 (in press).
}
\keyword{datasets}
