% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{getLibsizes2}
\alias{getLibsizes2}
\title{getLibsizes method for count Dat class, derived from getLibsizes1 method (countData class).}
\usage{
getLibsizes2(cD, subset = NULL, estimationType = c("quantile", "total",
  "edgeR"), quantile = 0.75, ...)
}
\arguments{
\item{cD}{A count Dat object.}

\item{subset}{Value}

\item{estimationType}{e.g. quantile, total, edgeR.}

\item{quantile}{A quantile, expressed as e.g. 0.75.}

\item{...}{Passthrough arguments.}
}
\description{
getLibsizes method for count Dat class, derived from getLibsizes1 method (countData class).
}
