% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibsizes3.R
\name{getLibsizes3}
\alias{getLibsizes3}
\title{getLibsizes3 Get Lib Sizes.}
\usage{
getLibsizes3(se, subset = NULL,
estimationType = c("quantile", "total",
"edgeR"),quantile = 0.75, ...)
}
\arguments{
\item{se}{An s.e. object.}

\item{subset}{Value}

\item{estimationType}{e.g. quantile, total, edgeR.}

\item{quantile}{A quantile, expressed as e.g. 0.75.}

\item{...}{Passthrough arguments.}
}
\value{
Libsize value
}
\description{
getLibsizes3 method for "SE" class object, derived
from getLibsizes2
}
\examples{
data(hmel.se)
getLibsizes3(se)
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
