% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{glSeq}
\alias{glSeq}
\title{glSeq LME4 wrapper.}
\usage{
glSeq(dm, model, ...)
}
\arguments{
\item{dm}{The dm data. Generally a reformatted countDat object from cD.DM.}

\item{model}{The model. Expressed using standard LME4 syntax, see vignette.}

\item{...}{passthrough arguments.}
}
\value{
Returns LME4 output.
}
\description{
This function is an LME4 wrapper for dosage analysis.
}
\details{
This function is an lme4 wrapper.
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
replicates(hm) <- reps
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
rpkm(hm) <- make_RPKM(hm)
f_hm <- quantFilter(hm, lo.bound = 0.4, hi.bound = 0.5)
dm <- cD.DM(f_hm)
glSeq(dm, "-1 + replicate")

}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
