% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmel.dat-data.R
\docType{data}
\name{hmel.dat}
\alias{hmel.dat}
\title{Hmel data set}
\format{An object of class \code{list} of length 3.}
\usage{
data("hmel.data.doser")
}
\description{
Hmel data set
}
\references{
The "doseR" package, 2018 (in press).
}
\keyword{datasets}
