% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{libsizes<-}
\alias{libsizes<-}
\alias{libsizes<-,countDat-method}
\alias{libsizes}
\alias{libsizes,countDat-method}
\title{libsizes method for testClass}
\usage{
libsizes(x) <- value

\S4method{libsizes}{countDat}(x) <- value

libsizes(x)

\S4method{libsizes}{countDat}(x)
}
\arguments{
\item{x}{Value}

\item{value}{Value}
}
\value{
libsize
}
\description{
libsizes method for testClass

libsizes method for testClass

libsizes method for testClass

libsizes method for testClass
}
\examples{
libsizes
}
\keyword{internal}
