% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RPKM.R
\name{make_RPKM}
\alias{make_RPKM}
\title{make_RPKM Make RPKM.}
\usage{
make_RPKM(cd_object)
}
\arguments{
\item{cd_object}{A countDat object.}
}
\value{
RPKM populated object
}
\description{
make_RPKM populates RPKM slot of modified countData (countDat)
S4 object.
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hmtr <- hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,replicates=reps,seglens=hmtr,
annotation=annotxn)
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
rpkm(hm) <- make_RPKM(hm)
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
