% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{nrow,countDat-method}
\alias{nrow,countDat-method}
\title{replicates method for testClass}
\usage{
\S4method{nrow}{countDat}(x)
}
\arguments{
\item{x}{Value}
}
\description{
replicates method for testClass
}
\keyword{internal}
