% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMedianReadCounts.R
\name{plotMedianReadCounts}
\alias{plotMedianReadCounts}
\title{countMedian Function to plot median read counts.}
\usage{
plotMedianReadCounts(G_RESULT3)
}
\arguments{
\item{G_RESULT3}{G_RESULT file.}
}
\value{
Returns median value plot for samples.
}
\description{
This function generates "median" read count plot.
}
\details{
This function plots median expression for selected sample.
}
\examples{
data(G.RESULT)
plotMedianReadCounts(G_RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
