\name{print.Absorbance}
\alias{print.Absorbance}
\alias{print}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print information regarding absorbance data object.
}
\description{
Prints information regarding drugs and cell lines to be analysed 
}
\usage{
\method{print}{Absorbance}(x, ...)
}
\arguments{\item{x}{A.data object created by the function \code{readDRdata}}
\item{...}{ignored}}