\name{print.bgModel}
\alias{print.bgModel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print information regarding absorbance data object.
}
\description{
Prints information regarding drugs and cell lines to be analysed 
}
\usage{
\method{print}{bgModel}(x, ...)
}
\arguments{\item{x}{A.data object created by the function bgModel}
\item{...}{ignored}}
