\name{print.bootstrap}
\alias{print.bootstrap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print information regarding absorbance data object.
}
\description{
Prints information regarding drugs and cell lines to be analysed 
}
\usage{
\method{print}{bootstrap}(x, ...)
}
\arguments{\item{x}{A.data object created by the function bootstrap}
\item{...}{ignored}}

