\name{print.drugColorCorrection}
\alias{print.drugColorCorrection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print information regarding absorbance data object.
}
\description{
Prints information regarding drugs and cell lines to be analysed 
}
\usage{
\method{print}{drugColorCorrection}(x, ...)
}
\arguments{\item{x}{A.data object created by the function drugColorCorrection}
\item{...}{ignored}}
