\name{print.summary.DRdata}
\alias{print.summary.DRdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print information regarding absorbance data object.
}
\description{
Prints information regarding drugs and cell lines to be analysed 
}
\usage{
\method{print}{summary.DRdata}(x, ...)
}
\arguments{\item{x}{A.data object created by the function summary.DRdata}
\item{...}{ignored}}
