% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratiosCorrelation.R
\name{ratiosCorrelation}
\alias{ratiosCorrelation}
\title{ratiosCorrelation Function to plot ratiosCorrelation.}
\usage{
ratiosCorrelation(GRESULT3)
}
\arguments{
\item{GRESULT3}{G_RESULT file.}
}
\value{
Returns ratiosCorrelation value plot for samples.
}
\description{
This function plots ratiosCorrelation.
}
\details{
This function plots ratiosCorrelation.
}
\examples{
data(G.RESULT)
ratiosCorrelation(G_RESULT)
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
