% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{replicates<-}
\alias{replicates<-}
\alias{replicates<-,countDat-method}
\alias{replicates}
\alias{replicates,countDat-method}
\title{replicates method for testClass}
\usage{
replicates(x) <- value

\S4method{replicates}{countDat}(x) <- value

replicates(x)

\S4method{replicates}{countDat}(x)
}
\arguments{
\item{x}{Value}

\item{value}{Value}

\item{x}{Value}

\item{x}{Value}
}
\value{
replicates populated object

x@replicates

x@replicates
}
\description{
replicates method for testClass

replicates method for testClass

replicates method for testClass

replicates method for testClass
}
\examples{

data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
replicates(hm) <- reps


data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
}
\keyword{internal}
