% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{rpkm<-}
\alias{rpkm<-}
\alias{rpkm<-,countDat-method}
\alias{rpkm}
\alias{rpkm,countDat-method}
\title{rpkm method for testClass}
\usage{
rpkm(x) <- value

\S4method{rpkm}{countDat}(x) <- value

rpkm(x)

\S4method{rpkm}{countDat}(x)
}
\arguments{
\item{x}{Value}

\item{value}{Value}

\item{x}{Value}

\item{value}{Value}

\item{x}{Value}

\item{x}{Value}
}
\value{
x@RPKM

x@RPKM
}
\description{
rpkm method for testClass

rpkm method for testClass

rpkm method for testClass

rpkm method for testClass
}
\examples{

data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" =
factor(hmel.dat$chromosome, levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)

data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
}
\keyword{internal}
