% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldMeanRatios.R
\name{scaffoldMeanRatios}
\alias{scaffoldMeanRatios}
\title{scaffoldMeanRatios Function to plot scaffoldMeanRatios.}
\usage{
scaffoldMeanRatios(GRESULT3)
}
\arguments{
\item{GRESULT3}{G_RESULT file.}
}
\value{
Returns scaffoldMeanRatios value plot for samples.
}
\description{
This function plots scaffoldMeanRatios.
}
\details{
This function plots scaffoldMeanRatios.
}
\references{
The "Zwyx" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
