% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-data.R
\docType{data}
\name{se}
\alias{se}
\title{Hmel data set}
\format{An object of class \code{SummarizedExperiment}
with 13619 rows and 6 columns.}
\usage{
data("hmel.se")
}
\description{
Hmel data set
}
\references{
To demonstrate a typical workflow using doseR, we will use a subset
of data from a study in _Heliconius_ butterflies.  _H. melpomene_
has 20 autosomes with chromosome 21 being the "Z" sex chromosome.
Like all Lepidoptera, this is a female-heterogametic species, so males
are diploid for the Z while females have a single Z and W chromosomes.
The relevant data are included in the doseR package, which must be installed.
These test data are unpublished. THIS IS THE SE VERSION OF THE SAMPLE DATA.
}
\keyword{datasets}
