% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleFilter.R
\name{simpleFilter}
\alias{simpleFilter}
\title{simpleFilter Function to filter expression data of a countDat object.}
\usage{
simpleFilter(cD, mean_cutoff=NULL, min_cutoff=NULL, median_cutoff=
NULL, counts=TRUE)
}
\arguments{
\item{cD}{A countDat object.}

\item{mean_cutoff}{The lower cutoff, using mean.}

\item{min_cutoff}{The cutoff, expressed as a minimum acceptable value.}

\item{median_cutoff}{The cutoff, using a median value.}

\item{counts}{Boolean, use raw counts data (default) or RPKM.}
}
\value{
Returns a filtered countDat object.
}
\description{
simpleFilter is a filtering function used to remove rows
(genes) of various expression data.
}
\details{
This function filters the expression of the supplied countDat
object, based on a selected cutoff.
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
replicates(hm) <- reps
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
rpkm(hm) <- make_RPKM(hm)
f_hm <- simpleFilter(hm, mean_cutoff=0.5)

}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
