% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{cD.DM}
\alias{cD.DM}
\title{cD.DM Function to convert count dat object to LME4 input.}
\usage{
cD.DM(cD, weightByLL = TRUE)
}
\arguments{
\item{cD}{A countDat object containing FPKM values and at least
one annotation column.}

\item{weightByLL}{Logical, weigh by log likelihood score.}
}
\value{
Returns LME4 input.
}
\description{
This function generates LME4 input.
}
\details{
This function converts count dat object.
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
replicates(hm) <- reps
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
rpkm(hm) <- make_RPKM(hm)
f_hm <- quantFilter(hm, lo.bound = 0.4, hi.bound = 0.5)
dm <- cD.DM(f_hm)

}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
