% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafs.R
\name{dafs}
\alias{dafs}
\title{dafs}
\usage{
dafs(VEC1, PLOT)
}
\arguments{
\item{VEC1}{Vector 1, not intended for user interaction.}

\item{PLOT}{Boolean, toggles plotting.}
}
\value{
Returns vx[which.min(vv)] to wrapper function.
}
\description{
This function filters the expression using DAFS (see ref).
This is a core function invoked by the DAFS wrapper.
}
\details{
This function filters the expression, using
Data Adaptive Flag method.
}
\examples{
library(mclust)
library(edgeR)
data(hmel.se)
f_se <- dafsFilter(se)

}
\references{
BMC Bioinformatics, 2014, 15:92
}
\author{
AJ Vaestermark, JR Walters.
}
