% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{getLibsizes2}
\alias{getLibsizes2}
\title{getLibsizes method for count Dat class, derived from getLibsizes1
method (countData class).}
\usage{
getLibsizes2(cD, subset = NULL,
estimationType = c("quantile", "total",
"edgeR"),quantile = 0.75, ...)
}
\arguments{
\item{cD}{A count Dat object.}

\item{subset}{Value}

\item{estimationType}{e.g. quantile, total, edgeR.}

\item{quantile}{A quantile, expressed as e.g. 0.75.}

\item{...}{Passthrough arguments.}
}
\value{
Libsize value
}
\description{
getLibsizes method for count Dat class, derived from getLibsizes1
method (countData class).
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,replicates=reps,
seglens=hm.tr,annotation=annotxn)
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
getLibsizes2(hm)

}
