% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmel.dat-data.R
\docType{data}
\name{hmel.dat}
\alias{hmel.dat}
\title{Hmel data set}
\format{An object of class \code{list} of length 3.}
\usage{
data("hmel.data.doser")
}
\description{
Hmel data set
}
\references{
To demonstrate a typical workflow using doseR, we will use
data from a study in Heliconius butterflies.  H. melpomene
has 20 autosomes with chromosome 21 being the "Z" sex chromosome.
Like all Lepidoptera, this is a female-heterogametic species, so males
are diploid for the Z while females have a single Z and W chromosomes.
The relevant data are included in the doseR package, which must be installed.
These test data are unpublished.
}
\keyword{datasets}
