% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RPKM.R
\name{make_RPKM}
\alias{make_RPKM}
\title{make_RPKM Make RPKM.}
\usage{
make_RPKM(cd_object)
}
\arguments{
\item{cd_object}{A countDat object.}
}
\description{
make_RPKM populates RPKM slot of modified countData (countDat) S4 object.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
