% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RPKM.R
\name{make_RPKM}
\alias{make_RPKM}
\title{make_RPKM Make RPKM.}
\usage{
make_RPKM(se)
}
\arguments{
\item{se}{An se object.}
}
\value{
RPKM populated object
}
\description{
make_RPKM populates RPKM slot of se
S4 object.
}
\examples{
data(hmel.se)
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
