% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantFilter.R
\name{quantFilter}
\alias{quantFilter}
\title{quantFilter Function to filter expression data of a
countDat object.}
\usage{
quantFilter (cD, lo.bound=.25, hi.bound=.75, MEDIAN = FALSE,
na.rm = TRUE)
}
\arguments{
\item{cD}{A countDat object.}

\item{lo.bound}{The lower cutoff, expressed as a percentage.}

\item{hi.bound}{The upper cutoff, expressed as a percentage.}

\item{MEDIAN}{Boolean, Calculate RowMeans or RowMedians.}

\item{na.rm}{Boolean, NA removal.}
}
\value{
Returns a filtered countDat object.
}
\description{
This function filters the expression of the supplied
countDat object; quantFilter is a filtering function used to remove
rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countDat
object, based on a selected percentage cutoff.
}
\examples{
data(hmel.data.doser)
reps <- c("Male", "Male", "Male", "Female", "Female", "Female")
annotxn <- data.frame("Chromosome" = factor(hmel.dat$chromosome,
levels = 1:21))
hm.tr<-hmel.dat$trxLength
hm<-new("countDat",data=hmel.dat$readcounts,seglens=hm.tr,
annotation=annotxn)
replicates(hm) <- reps
libsizes(hm) <- getLibsizes2(hm, estimationType = "total")
rpkm(hm) <- make_RPKM(hm)
f_hm <- quantFilter(hm, lo.bound=0.5)

}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
