% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDat.R
\docType{methods}
\name{replicates<-}
\alias{replicates<-}
\alias{replicates<-,countDat-method}
\alias{groups<-}
\alias{groups<-,countDat-method}
\alias{groups}
\alias{groups,countDat-method}
\title{replicates method for testClass}
\usage{
replicates(x) <- value

\S4method{replicates}{countDat}(x) <- value

groups(x) <- value

\S4method{groups}{countDat}(x) <- value

groups(x)

\S4method{groups}{countDat}(x)
}
\arguments{
\item{x}{Value}

\item{value}{Value}

\item{x}{Value}

\item{value}{Value}

\item{x}{Value}

\item{value}{Value}
}
\description{
replicates method for testClass

replicates method for testClass

replicates method for testClass

replicates method for testClass

replicates method for testClass

replicates method for testClass
}
\keyword{internal}
