% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleFilter.R
\name{simpleFilter}
\alias{simpleFilter}
\title{simpleFilter Function to filter expression data of a countDat object.}
\usage{
simpleFilter(cD, mean_cutoff=NULL, min_cutoff=NULL, median_cutoff=NULL, counts=TRUE)
}
\arguments{
\item{cD}{A countDat object.}

\item{mean_cutoff}{The lower cutoff, using mean.}

\item{min_cutoff}{The cutoff, expressed as a minimum acceptable value.}

\item{median_cutoff}{The cutoff, using a median value.}

\item{counts}{Boolean, use raw counts data (default) or RPKM.}
}
\value{
Returns a filtered countDat object.
}
\description{
simpleFilter is a filtering function used to remove rows (genes) of various expression data.
}
\details{
This function filters the expression of the supplied countDat object, based on a selected cutoff.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
