% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diffs.R
\name{test_diffs}
\alias{test_diffs}
\title{test_diffs Statistics function to summarize expression data of a countDat object, using ratios between selected treatments.}
\usage{
test_diffs(cD, groupings= NULL, treatment1=NULL, treatment2=NULL,
                  mode_mean=TRUE, LOG2=TRUE)
}
\arguments{
\item{cD}{A countDat object.}

\item{groupings}{A grouping (annotation column), e.g. groupings="something".}

\item{treatment1}{Symbol, treatment 1.}

\item{treatment2}{Symbol, treatment 2.}

\item{mode_mean}{Boolean, Calculate RowMeans or RowMedians.}

\item{LOG2}{Boolean, Calculate LOG2.}
}
\value{
Returns an invisible list of summary statistics, kruskal test and raw data of a countDat object, using ratios between selected treatments.
}
\description{
generateStats is a summary function used on various expression data, using ratios between selected treatments.
}
\details{
This function completes summary statistics of the expression of the supplied countDat object.
}
\references{
The "doseR" package, 2018 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
